src/commands/ImportCommand.js -> commands/import/index.js
