+++ commands/import/package.json
@@ -14,7 +14,8 @@
   },
   "files": [
     "command.js",
-    "index.js"
+    "index.js",
+    "lib"
   ],
   "main": "index.js",
   "engines": {
@@ -30,5 +31,15 @@
   },
   "scripts": {
     "test": "echo \"Run tests from root\" && exit 1"
+  },
+  "dependencies": {
+    "@lerna/child-process": "file:../../core/child-process",
+    "@lerna/command": "file:../../core/command",
+    "@lerna/fs-utils": "file:../../core/fs-utils",
+    "@lerna/git-utils": "file:../../core/git-utils",
+    "@lerna/prompt": "file:../../core/prompt",
+    "@lerna/validation-error": "file:../../core/validation-error",
+    "dedent": "^0.7.0",
+    "p-map-series": "^1.0.0"
   }
 }
